# Conflicts:
#	Integration/Scripts/IC/uspICDCUomMigrationPt.sql
#	i21Database/Scripts/Post-Deployment/PostDeployment.sql
#	i21Database/Scripts/Post-Deployment/SM/DefaultData/4_StartingNumbers.sql
#	i21Database/dbo/Functions/fnRKGetOpenContractHistory.sql
#	i21Database/dbo/Functions/fnRKGetOpenFutureByDate.sql
#	i21Database/dbo/Stored Procedures/uspAPPostBill.sql
#	i21Database/dbo/Stored Procedures/uspAPUpdateVoucherPayableQty.sql
#	i21Database/dbo/Stored Procedures/uspAPUpdateVoucherTotal.sql
#	i21Database/dbo/Stored Procedures/uspARAddItemToInvoice.sql
#	i21Database/dbo/Stored Procedures/uspARAddItemToInvoices.sql
#	i21Database/dbo/Stored Procedures/uspARCreateInvoiceFromShipment.sql
#	i21Database/dbo/Stored Procedures/uspARCustomerStatementFullDetailReport.sql
#	i21Database/dbo/Stored Procedures/uspARInvoiceReport.sql
#	i21Database/dbo/Stored Procedures/uspARPopulateInvalidPostInvoiceData.sql
#	i21Database/dbo/Stored Procedures/uspARPopulateInvalidPostPaymentData.sql
#	i21Database/dbo/Stored Procedures/uspARPostInvoice.sql
#	i21Database/dbo/Stored Procedures/uspARPostInvoiceNew.sql
#	i21Database/dbo/Stored Procedures/uspARUpdateOverageContracts.sql
#	i21Database/dbo/Stored Procedures/uspCTCreateDetailHistory.sql
#	i21Database/dbo/Stored Procedures/uspCTGetContractBalance.sql
#	i21Database/dbo/Stored Procedures/uspCTLogSummary.sql
#	i21Database/dbo/Stored Procedures/uspICRebuildInventoryValuation.sql
#	i21Database/dbo/Stored Procedures/uspLGPostWeightClaims.sql
#	i21Database/dbo/Stored Procedures/uspMFBatchSheetReportDetail.sql
#	i21Database/dbo/Stored Procedures/uspMFGenerateDemand.sql
#	i21Database/dbo/Stored Procedures/uspRKGetInHouse.sql
#	i21Database/dbo/Stored Procedures/uspTFProcessBeforePreview.sql
#	i21Database/dbo/Tables/tblARInvoiceReportStagingTable.sql
#	i21Database/dbo/Tables/tblSMCompanyPreference.sql
#	i21Database/dbo/Views/vyuAPPayablesAgingDeletedForeign.sql
#	i21Database/dbo/Views/vyuAPPayablesForeign.sql
#	i21Database/dbo/Views/vyuARCustomerSearch.sql
#	i21Database/dbo/Views/vyuICGetItemStock.sql
#	i21Database/dbo/Views/vyuSCTicketView.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/User Defined Types/StagingRadiantFGM.sql
@@ -1,6 +1,6 @@
 CREATE TYPE [dbo].[StagingRadiantFGM] AS TABLE(
 	[intRowCount] [int] NULL,
-	[intFuelGradeID] [int] NULL,
+	[intFuelGradeID] [bigint] NULL,
 	[dblFuelGradeNonResettableTotalVolume] [numeric](18, 6) NULL,
 	[dblFuelGradeNonResettableTotalAmount] [numeric](18, 6) NULL,
 	[dblFuelGradeSalesVolume] [numeric](18, 6) NULL,
